Extract progress tracking code to separate module
